se creo la vista home independiente del registro
