+++ src/index.html
@@ -14,6 +14,7 @@
     <nav>
         <ul>
             <li><a href="#/Home">Home</a></li>
+            <li><a href="#/Registro">Registro</a></li>
             <li><a href="#/login">Iniciar Sesion</a></li>
             <li><a href="#/logout" id="logout">Cerrar Sesion</a></li>
         </ul>
